+++ trtools/utils/tests/test_trharmonizer.py
@@ -61,9 +61,9 @@ def get_dummy_record():
     )
 
 triploid_gts = np.array([
-    [0, 0, -1],
-    [0, 0, -1],
-    [0, 0, -1],
+    [0, 0, -2],
+    [0, 0, -2],
+    [0, 0, -2],
     [0, 0, 0]
 ])
 def get_triploid_record():
@@ -272,7 +272,7 @@ def test_TRRecord_GetGenotypes():
                 [alt_alleles[0], alt_alleles[0]],
                 [alt_alleles[0], alt_alleles[1]],
                 [alt_alleles[1], alt_alleles[1]],
-                [ref_allele, '']]
+                [ref_allele, '.']]
     true_gts = np.array(true_gts)
     true_len_gts = np.array([[3, 4], [4, 4], [4, 4], [4, 6], [6, 6], [3, -1]])
     assert np.all(rec.GetGenotypeIndicies()[:, :-1] ==
@@ -283,112 +283,107 @@ def test_TRRecord_GetGenotypes():
     # Test example where alt=[]
     triploid_record = get_triploid_record()
     rec = trh.TRRecord(triploid_record, ref_allele, [], "CAG", "", None)
-    called_indicies = triploid_gts != -1
-    called_indicies = np.concatenate((called_indicies,
-                                      np.zeros((4, 1), dtype=bool)),
-                                     axis=1)
-    assert np.all(rec.GetGenotypeIndicies()[called_indicies] == 0)
-    assert np.all(rec.GetLengthGenotypes()[called_indicies] == 3)
-    assert np.all(rec.GetStringGenotypes()[called_indicies] == ref_allele)
+    true_len_gts = [[3, 3, -2],
+                [3, 3, -2],
+                [3, 3, -2],
+                [3, 3, 3]]
+    true_len_gts = np.array(true_len_gts)
+    true_idx_gts = [[0, 0, -2],
+                    [0, 0, -2],
+                    [0, 0, -2],
+                    [0, 0, 0]]
+    true_idx_gts = np.array(true_idx_gts)
+    true_gts = [[ref_allele, ref_allele, ','],
+                [ref_allele, ref_allele, ','],
+                [ref_allele, ref_allele, ','],
+                [ref_allele, ref_allele, ref_allele]]
+    true_gts = np.array(true_gts)
+
+    assert np.all(rec.GetGenotypeIndicies()[:, :-1] == true_idx_gts)
+    assert np.all(rec.GetLengthGenotypes()[:, :-1] == true_len_gts)
+    assert np.all(rec.GetStringGenotypes()[:, :-1] == true_gts)
 
     # Test example with fewer alt_alleles than the max genotype index
     with pytest.raises(ValueError):
         trh.TRRecord(dummy_record, ref_allele, [], "CAG", "", None)
 
 
+def _dicts_equal(dict1, dict2):
+    return (all(k in dict2 and v == dict2[k] for k,v in dict1.items()) and
+            len(dict1) == len(dict2))
+
 def test_GetGenotypeCounts():
-    # Test working example, no samplelist
+    # Test working example, no sample_index,
+    # also a call with a missing haplotype
     dummy_record = get_dummy_record()
     ref_allele = dummy_record.REF
     alt_alleles = dummy_record.ALT
     rec = trh.TRRecord(dummy_record, ref_allele, alt_alleles, "CAG", "", None)
-    true_idx_counts = {(0, 1) : 1,
-                       (1, 1) : 2,
-                       (1, 2) : 1,
-                       (2, 2) : 1,
-                       (-1, 0) : 1}
+    true_idx_gt_counts = {(0, 1) : 1,
+                          (1, 1) : 2,
+                          (1, 2) : 1,
+                          (2, 2) : 1}
     true_gt_counts = {(ref_allele, alt_alleles[0]): 1,
                       (alt_alleles[0], alt_alleles[0]): 2,
                       (alt_alleles[0], alt_alleles[1]): 1,
-                      (alt_alleles[1], alt_alleles[1]): 1,
-                      ('', ref_allele): 1}
+                      (alt_alleles[1], alt_alleles[1]): 1}
     true_len_gt_counts = {
-        (3.0, 4.0): 1,
-        (4.0, 4.0): 2,
-        (4.0, 6.0): 1,
-        (6.0, 6.0): 1,
-        (-1.0, 3.0): 1
+        (3, 4): 1,
+        (4, 4): 2,
+        (4, 6): 1,
+        (6, 6): 1
     }
 
     gt_counts_uselength = rec.GetGenotypeCounts()
     gt_counts_nolength = rec.GetGenotypeCounts(uselength=False)
     gt_counts_idx = rec.GetGenotypeCounts(index=True)
-    assert (all(
-        v == true_idx_counts[k] for k, v in gt_counts_idx.items()
-    ) and len(gt_counts_idx) == len(true_idx_counts))
-    assert (all(
-        v == true_len_gt_counts[k] for k, v in gt_counts_uselength.items()
-    ) and len(gt_counts_uselength) == len(true_len_gt_counts))
-    assert (all(
-        v == true_gt_counts[k] for k, v in gt_counts_nolength.items()
-    ) and len(gt_counts_nolength) == len(true_gt_counts))
+    assert _dicts_equal(true_idx_gt_counts, gt_counts_idx)
+    assert _dicts_equal(true_len_gt_counts, gt_counts_uselength)
+    assert _dicts_equal(true_gt_counts, gt_counts_nolength)
 
     # Test example where alt=[]
     triploid_record = get_triploid_record()
     rec = trh.TRRecord(triploid_record, ref_allele, [], "CAG", "", None)
-    true_idx_gt_counts = {(0, 0, 0): 1, (-1, 0, 0): 3}
-    true_len_gt_counts = {(3.0, 3.0, 3.0): 1, (-1.0, 3.0, 3.0): 3}
+    true_idx_gt_counts = {(0, 0, 0): 1, (-2, 0, 0): 3}
+    true_len_gt_counts = {(3, 3, 3): 1, (-2, 3, 3): 3}
     true_gt_counts = {(ref_allele, ref_allele, ref_allele): 1,
-                      ('', ref_allele, ref_allele): 3}
+                      (',', ref_allele, ref_allele): 3}
     gt_counts_idx = rec.GetGenotypeCounts(index=True)
     gt_counts_uselength = rec.GetGenotypeCounts()
     gt_counts_nolength = rec.GetGenotypeCounts(uselength=False)
-    assert (all(
-        v == true_idx_gt_counts[k] for k, v in gt_counts_idx.items()
-    ) and len(gt_counts_idx) == len(true_idx_gt_counts))
-    assert (all(
-        v == true_len_gt_counts[k] for k, v in gt_counts_uselength.items()
-    ) and len(gt_counts_uselength) == len(true_len_gt_counts))
-    assert (all(
-        v == true_gt_counts[k] for k, v in gt_counts_nolength.items()
-    ) and len(gt_counts_nolength) == len(true_gt_counts))
+    assert _dicts_equal(true_idx_gt_counts, gt_counts_idx)
+    assert _dicts_equal(true_len_gt_counts, gt_counts_uselength)
+    assert _dicts_equal(true_gt_counts, gt_counts_nolength)
 
     # Test example where there are no samples
     rec = trh.TRRecord(get_nocall_record(), ref_allele, [], "CAG", "", None)
     # expect a entry for counting no-calls
-    assert len(rec.GetGenotypeCounts(index=True)) == 1
-    assert len(rec.GetGenotypeCounts()) == 1
-    assert len(rec.GetGenotypeCounts(uselength=True)) == 1
+    assert len(rec.GetGenotypeCounts(index=True)) == 0
+    assert len(rec.GetGenotypeCounts()) == 0
+    assert len(rec.GetGenotypeCounts(uselength=True)) == 0
+    assert len(rec.GetGenotypeCounts(index=True, include_nocalls=True)) == 1
+    assert len(rec.GetGenotypeCounts(include_nocalls=True)) == 1
+    assert len(rec.GetGenotypeCounts(uselength=True, include_nocalls=True)) == 1
 
-    # Test working example with samplelist
-    true_idx_gt_counts_slist = {(0, 1): 1, (1, 1): 1, (-1, 0): 1}
-    true_gt_counts_slist = {(ref_allele, alt_alleles[0]): 1,
-                            (alt_alleles[0], alt_alleles[0]): 1,
-                            ('', ref_allele): 1}
-    true_len_gt_counts_slist = {(3, 4): 1, (4, 4): 1, (-1, 3): 1}
-    slist = [0, 2, 5]
+    # Test working example with sample_index
+    true_idx_gt_counts_sindex = {(0, 1): 1, (1, 1): 1}
+    true_gt_counts_sindex = {(ref_allele, alt_alleles[0]): 1,
+                            (alt_alleles[0], alt_alleles[0]): 1}
+    true_len_gt_counts_sindex = {(3, 4): 1, (4, 4): 1}
+    sindex = [0, 2, 5]
     dummy_record = get_dummy_record()
     rec = trh.TRRecord(dummy_record, ref_allele, alt_alleles, "CAG", "", None)
-    gt_counts_idx_slist = rec.GetGenotypeCounts(samplelist=slist, index=True)
-    gt_counts_uselength_slist = rec.GetGenotypeCounts(samplelist=slist)
-    gt_counts_nolength_slist = rec.GetGenotypeCounts(samplelist=slist,
+    gt_counts_idx_sindex = rec.GetGenotypeCounts(sample_index=sindex, index=True)
+    gt_counts_uselength_sindex = rec.GetGenotypeCounts(sample_index=sindex)
+    gt_counts_nolength_sindex = rec.GetGenotypeCounts(sample_index=sindex,
                                                      uselength=False)
-    assert (all(
-        v == true_idx_gt_counts_slist[k] for k, v in
-        gt_counts_idx_slist.items()
-    ) and len(gt_counts_idx_slist) == len(true_idx_gt_counts_slist))
-    assert (all(
-        v == true_len_gt_counts_slist[k] for k, v in
-        gt_counts_uselength_slist.items()
-    ) and len(gt_counts_uselength_slist) == len(true_len_gt_counts_slist))
-    assert (all(
-        v == true_gt_counts_slist[k] for k, v
-        in gt_counts_nolength_slist.items()
-    ) and len(gt_counts_nolength_slist) == len(true_gt_counts_slist))
+    assert _dicts_equal(true_idx_gt_counts_sindex, gt_counts_idx_sindex)
+    assert _dicts_equal(true_len_gt_counts_sindex, gt_counts_uselength_sindex)
+    assert _dicts_equal(true_gt_counts_sindex, gt_counts_nolength_sindex)
 
 
 def test_GetAlleleCounts():
-    # Test working example, no samplelist
+    # Test working example, no sample_index
     dummy_record = get_dummy_record()
     ref_allele = dummy_record.REF
     alt_alleles = dummy_record.ALT
@@ -400,15 +395,9 @@ def test_GetAlleleCounts():
     al_counts_uselength = rec.GetAlleleCounts()
     al_counts_nolength = rec.GetAlleleCounts(uselength=False)
     al_counts_idx = rec.GetAlleleCounts(index=True)
-    assert (all(
-        v == true_idx_al_counts[k] for k, v in al_counts_idx.items()
-    ) and len(al_counts_idx) == len(true_idx_al_counts))
-    assert (all(
-        v == true_len_al_counts[k] for k, v in al_counts_uselength.items()
-    ) and len(al_counts_uselength) == len(true_len_al_counts))
-    assert (all(
-        v == true_al_counts[k] for k, v in al_counts_nolength.items()
-    ) and len(al_counts_nolength) == len(true_al_counts))
+    assert _dicts_equal(true_idx_al_counts, al_counts_idx)
+    assert _dicts_equal(true_len_al_counts, al_counts_uselength)
+    assert _dicts_equal(true_al_counts, al_counts_nolength)
 
     # Test example where alt=[]
     rec = trh.TRRecord(get_triploid_record(), ref_allele, [], "CAG", "", None)
@@ -418,15 +407,9 @@ def test_GetAlleleCounts():
     al_counts_idx = rec.GetAlleleCounts(index=True)
     al_counts_uselength = rec.GetAlleleCounts()
     al_counts_nolength = rec.GetAlleleCounts(uselength=False)
-    assert (all(
-        v == true_idx_al_counts[k] for k, v in al_counts_idx.items()
-    ) and len(al_counts_idx) == len(true_idx_al_counts))
-    assert (all(
-        v == true_len_al_counts[k] for k, v in al_counts_uselength.items()
-    ) and len(al_counts_uselength) == len(true_len_al_counts))
-    assert (all(
-        v == true_al_counts[k] for k, v in al_counts_nolength.items()
-    ) and len(al_counts_nolength) == len(true_al_counts))
+    assert _dicts_equal(true_idx_al_counts, al_counts_idx)
+    assert _dicts_equal(true_len_al_counts, al_counts_uselength)
+    assert _dicts_equal(true_al_counts, al_counts_nolength)
 
     # Test example where there are no samples
     rec = trh.TRRecord(get_nocall_record(), ref_allele, [], "CAG", "", None)
@@ -434,32 +417,23 @@ def test_GetAlleleCounts():
     assert len(rec.GetAlleleCounts()) == 0
     assert len(rec.GetAlleleCounts(uselength=True)) == 0
 
-    # Test working example with samplelist
-    true_idx_al_counts_slist = {0: 2, 1: 3}
-    true_len_al_counts_slist = {3: 2, 4: 3}
-    true_al_counts_slist = {ref_allele: 2, alt_alleles[0]: 3}
-    slist = [0, 2, 5]
+    # Test working example with sample_index
+    true_idx_al_counts_sindex = {0: 2, 1: 3}
+    true_len_al_counts_sindex = {3: 2, 4: 3}
+    true_al_counts_sindex = {ref_allele: 2, alt_alleles[0]: 3}
+    sindex = [0, 2, 5]
     rec = trh.TRRecord(dummy_record, ref_allele, alt_alleles, "CAG", "", None)
-    al_counts_idx_slist = rec.GetAlleleCounts(samplelist=slist, index=True)
-    al_counts_uselength_slist = rec.GetAlleleCounts(samplelist=slist)
-    al_counts_nolength_slist = rec.GetAlleleCounts(samplelist=slist,
+    al_counts_idx_sindex = rec.GetAlleleCounts(sample_index=sindex, index=True)
+    al_counts_uselength_sindex = rec.GetAlleleCounts(sample_index=sindex)
+    al_counts_nolength_sindex = rec.GetAlleleCounts(sample_index=sindex,
                                                    uselength=False)
-    assert (all(
-        v == true_idx_al_counts_slist[k] for k, v in
-        al_counts_idx_slist.items()
-    ) and len(al_counts_idx_slist) == len(true_idx_al_counts_slist))
-    assert (all(
-        v == true_len_al_counts_slist[k] for k, v in
-        al_counts_uselength_slist.items()
-    ) and len(al_counts_uselength_slist) == len(true_len_al_counts_slist))
-    assert (all(
-        v == true_al_counts_slist[k] for k, v
-        in al_counts_nolength_slist.items()
-    ) and len(al_counts_nolength_slist) == len(true_al_counts_slist))
+    assert _dicts_equal(true_idx_al_counts_sindex, al_counts_idx_sindex)
+    assert _dicts_equal(true_len_al_counts_sindex, al_counts_uselength_sindex)
+    assert _dicts_equal(true_al_counts_sindex, al_counts_nolength_sindex)
 
 
 def test_GetAlleleFreqs():
-    # Test working example, no samplelist
+    # Test working example, no sample_index
     dummy_record = get_dummy_record()
     ref_allele = dummy_record.REF
     alt_alleles = dummy_record.ALT
@@ -509,28 +483,28 @@ def test_GetAlleleFreqs():
     assert len(rec.GetAlleleFreqs()) == 0
     assert len(rec.GetAlleleFreqs(uselength=True)) == 0
 
-    # Test working example with samplelist
-    true_idx_al_freqs_slist = {0: 0.4, 1: 0.6}
-    true_len_al_freqs_slist = {3: 0.4, 4: 0.6}
-    true_al_freqs_slist = {ref_allele: 0.4, alt_alleles[0]: 0.6}
-    slist = [0, 2, 5]
+    # Test working example with sample_index
+    true_idx_al_freqs_sindex = {0: 0.4, 1: 0.6}
+    true_len_al_freqs_sindex = {3: 0.4, 4: 0.6}
+    true_al_freqs_sindex = {ref_allele: 0.4, alt_alleles[0]: 0.6}
+    sindex = [0, 2, 5]
     rec = trh.TRRecord(dummy_record, ref_allele, alt_alleles, "CAG", "", None)
-    al_freqs_idx_slist = rec.GetAlleleFreqs(samplelist=slist, index=True)
-    al_freqs_uselength_slist = rec.GetAlleleFreqs(samplelist=slist)
-    al_freqs_nolength_slist = rec.GetAlleleFreqs(samplelist=slist,
+    al_freqs_idx_sindex = rec.GetAlleleFreqs(sample_index=sindex, index=True)
+    al_freqs_uselength_sindex = rec.GetAlleleFreqs(sample_index=sindex)
+    al_freqs_nolength_sindex = rec.GetAlleleFreqs(sample_index=sindex,
                                                  uselength=False)
     assert (all(
-        v == true_idx_al_freqs_slist[k] for k, v in
-        al_freqs_idx_slist.items()
-    ) and len(al_freqs_idx_slist) == len(true_idx_al_freqs_slist))
+        v == true_idx_al_freqs_sindex[k] for k, v in
+        al_freqs_idx_sindex.items()
+    ) and len(al_freqs_idx_sindex) == len(true_idx_al_freqs_sindex))
     assert (all(
-        v == true_len_al_freqs_slist[k] for k, v in
-        al_freqs_uselength_slist.items()
-    ) and len(al_freqs_uselength_slist) == len(true_len_al_freqs_slist))
+        v == true_len_al_freqs_sindex[k] for k, v in
+        al_freqs_uselength_sindex.items()
+    ) and len(al_freqs_uselength_sindex) == len(true_len_al_freqs_sindex))
     assert (all(
-        v == true_al_freqs_slist[k] for k, v
-        in al_freqs_nolength_slist.items()
-    ) and len(al_freqs_nolength_slist) == len(true_al_freqs_slist))
+        v == true_al_freqs_sindex[k] for k, v
+        in al_freqs_nolength_sindex.items()
+    ) and len(al_freqs_nolength_sindex) == len(true_al_freqs_sindex))
 
 
 def test_GetMaxAllele():
@@ -539,13 +513,13 @@ def test_GetMaxAllele():
     ref_allele = dummy_record.REF
     alt_alleles = dummy_record.ALT
     rec = trh.TRRecord(dummy_record, ref_allele, alt_alleles, "CAG", "", None)
-    true_al_max = 6.0
+    true_al_max = 6
     al_max = rec.GetMaxAllele()
     assert al_max == true_al_max
 
     # Test example where alt=[]
     rec = trh.TRRecord(get_triploid_record(), ref_allele, [], "CAG", "", None)
-    true_al_max = 3.0
+    true_al_max = 3
     al_max = rec.GetMaxAllele()
     assert al_max == true_al_max
 
@@ -555,12 +529,12 @@ def test_GetMaxAllele():
     al_max = rec.GetMaxAllele()
     assert np.isnan(al_max)
 
-    # Test working example with samplelist
-    slist = [0, 2, 5]
-    true_al_max_slist = 4.0
+    # Test working example with sample_index
+    sindex = [0, 2, 5]
+    true_al_max_sindex = 4
     rec = trh.TRRecord(dummy_record, ref_allele, alt_alleles, "CAG", "", None)
-    al_max_slist = rec.GetMaxAllele(samplelist=slist)
-    assert al_max_slist == true_al_max_slist
+    al_max_sindex = rec.GetMaxAllele(sample_index=sindex)
+    assert al_max_sindex == true_al_max_sindex
 
 
 #### Test TRRecordHarmonizer on different files ####
@@ -756,6 +730,8 @@ def test_HarmonizeRecord(vcfdir):
     assert (tr_rec3.alt_alleles ==
             ['tgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtgtg'.upper()])
     assert tr_rec3.motif == 'tg'.upper()
+    assert tr_rec3.ref_allele_length == 26
+    assert tr_rec3.alt_allele_lengths == [24]
 
     # hipstr
     hipstr_trh = trh.TRRecordHarmonizer(hipstr_vcf)